ctmap: refactor cached GC interval out of ctmap package.
[ upstream commit 44834034ee3a2e12f44fc629a644a7640e66d5f2 ]

In subsequent commits we will want to make decisions on when
to set the cached GC interval depending on whether both enabled
ip families have had their GC run or just one (of two).

To do this, we need to move the cachedGCInterval out of the ctmap
package and make this explicit in the GetInterval function.

To prep for that we fix the tests and set the cached interval
outside of GetInterval.

Addresses: #40820

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@cilium.io>